pyproject.toml
[tool.poetry]
name = "serverless-mailer"
version = "0.1.0"
description = ""
authors = ["Mayank Chauhan <mayank.chauhan15@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.7"
boto3 = "^1.9"
python-dotenv = "^0.10.1"
lambda-decorators = "^0.2.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
